systemd-sleep: add a check of a partition size before a choose location to hibernation
If there are several partitions of the swap in the system,
and they have  different sizes and prioritys,
the behavior of the function "find_hibernate_location"
depend of an order of partition in the list "/proc/swaps".

It is happen because the function "find_hibernate_location" prefer
a partition with highest priority and doesn't check
a partition size with lower priority.